+++ src/missions_lib/__init__.py
@@ -1,10 +1,5 @@
 # __init__.py
-#from .missions_lib import TakeAMessage
-from .greet_all import Greeting
-from .greet_all import PrepareMovementGeeting
-from .greet_all import GreetingHelper
+from .greet_all import Mission2StateMachine
 
-from .object_search import PrepareMovementObjectSearch
-from .object_search import ChoiceObjectSearch
-from .object_search import ObjectSearchHelper
+