+++ src/rodney_missions_node.py
@@ -5,14 +5,12 @@ import sys
 import rospy
 import missions_lib
 from std_msgs.msg import String, Empty
-from servo_msgs.msg import pan_tilt
-from face_recognition_msgs.msg import face_recognition
-from tf_object_detection.msg import detection_results
 from smach import State, StateMachine
 from smach_ros import MonitorState, SimpleActionState, IntrospectionServer
 from actionlib_msgs.msg import GoalStatus
 from speech.msg import voice
 from head_control.msg import point_headAction, point_headGoal
+from math import pi
 
 # Helper class to hold code used by serveral different states
 class MissionsHelper():
@@ -22,27 +20,26 @@ class MissionsHelper():
         
         # Obtain values from the parameter server
         # Minium/Maximum range movment of camera
-        self.__pan_min = rospy.get_param("/servo/index0/pan_min", 0)
-        self.__pan_max = rospy.get_param("/servo/index0/pan_max", 180)
-        self.__tilt_min = rospy.get_param("/servo/index0/tilt_min", 0)
-        self.__tilt_max = rospy.get_param("/servo/index0/tilt_max", 180);
+        self.__pan_min = rospy.get_param("/servo/index0/pan/min", -(pi/2.0))
+        self.__pan_max = rospy.get_param("/servo/index0/pan/max", pi/2.0)
+        self.__tilt_min = rospy.get_param("/servo/index0/tilt/min", -(pi/2.0))
+        self.__tilt_max = rospy.get_param("/servo/index0/tilt/max", pi/2.0);
         # Default position after mission ends
-        self.__camera_default_pan_position = rospy.get_param("/head/position/pan", 90)
-        self.__camera_default_tilt_position = rospy.get_param("/head/position/tilt", 45)
+        self.__camera_default_pan_position = rospy.get_param("/head/position/pan", 0.0)
+        self.__camera_default_tilt_position = rospy.get_param("/head/position/tilt", 0.0)
         # Step value to move the camera by when searching
-        self.__pan_step_value = rospy.get_param("/head/view_step/pan", 25)
-        self.__tilt_step_value = rospy.get_param("/head/view_step/tilt", 25)
+        self.__pan_step_value = rospy.get_param("/head/view_step/pan", 0.436332)
+        self.__tilt_step_value = rospy.get_param("/head/view_step/tilt", 0.436332)
         # Step value to move the camera in manual mode
-        self.__manual_pan_step_value = rospy.get_param("/head/maunal_view_step/pan", 10)
-        self.__manual_tilt_step_value = rospy.get_param("/head/maunal_view_step/tilt", 10)
+        self.__manual_pan_step_value = rospy.get_param("/head/manual_view_step/pan", 0.174533)
+        self.__manual_tilt_step_value = rospy.get_param("/head/manual_view_step/tilt", 0.174533)
         
         # When true and scanning pan angle will increase, otherwise decrease
         self.__increase_pan = True
         
         # Position that will be requested to move the head/camera to
-        self.__position_request = pan_tilt()
-        self.__position_request.pan = self.__camera_default_pan_position   
-        self.__position_request.tilt = self.__camera_default_tilt_position
+        self.__position_request_pan = self.__camera_default_pan_position   
+        self.__position_request_tilt = self.__camera_default_tilt_position
  
     def Speak(self, text_to_speak, text_to_display):
         voice_msg = voice()
@@ -65,22 +62,21 @@ class MissionsHelper():
     # Function to return the camera start position when scanning within head movement range         
     def CameraToStartPos(self):
         # Set the camera position to pan min and tilt min
-        self.__position_request.pan = self.__pan_min
-        self.__position_request.tilt = self.__tilt_min
+        self.__position_request_pan = self.__pan_min
+        self.__position_request_tilt = self.__tilt_max
         # Set the variable that says which direction the pan is going. Start by incrementing
         self.__increase_pan_ = True
-        return self.__position_request
+        return self.__position_request_pan, self.__position_request_tilt 
 
     # Function to keep track of position after action to set to default position
     def CameraAtDefaultPos(self, userdata, status, result):
         if status == GoalStatus.SUCCEEDED:
-            self.__position_request.pan = self.__camera_default_pan_position
-            self.__position_request.tilt = self.__camera_default_tilt_position
+            self.__position_request_pan = self.__camera_default_pan_position
+            self.__position_request_tilt = self.__camera_default_tilt_position
         
     # Function returns camera default position
     def CameraDefaultPos(self):
-        default_position = pan_tilt(self.__camera_default_pan_position, self.__camera_default_tilt_position)
-        return default_position           
+        return self.__camera_default_pan_position, self.__camera_default_tilt_position           
  
     # Function to return the next position when scanning within the head movement range.
     # Also returns indication if all areas scanned or more left           
@@ -88,91 +84,92 @@ class MissionsHelper():
         all_areas_scanned = False
         # Calculate the next position of the head/camera        
         if self.__increase_pan == True:
-            if self.__position_request.pan == self.__pan_max:
+            if self.__position_request_pan == self.__pan_max:
                 # Last scan was at the edge, move tilt up and then pan the other way
-                self.__increase_pan = False                    
-                self.__position_request.tilt += self.__tilt_step_value
+                self.__increase_pan = False
+                self.__position_request_tilt -= self.__tilt_step_value
                   
-                if self.__position_request.tilt > self.__tilt_max:
+                if self.__position_request_tilt < self.__tilt_min:
                     all_areas_scanned = True
             else:
-                self.__position_request.pan += self.__pan_step_value
+                self.__position_request_pan += self.__pan_step_value
                     
-                if self.__position_request.pan > self.__pan_max:
+                if self.__position_request_pan > self.__pan_max:
                     # Moved out of range, put back on max
-                    self.__position_request.pan = self.__pan_max
+                    self.__position_request_pan = self.__pan_max
         else:    
-            if self.__position_request.pan == self.__pan_min:
+            if self.__position_request_pan == self.__pan_min:
                 # Last scan was at the edge, move tilt up and then pan the other way
                 self.__increase_pan = True
-                self.__position_request.tilt += self.__tilt_step_value
+                self.__position_request_tilt -= self.__tilt_step_value
                     
-                if self.__position_request.tilt > self.__tilt_max:
+                if self.__position_request_tilt < self.__tilt_min:
                     all_areas_scanned = True
             else:
-                self.__position_request.pan -= self.__pan_step_value
+                self.__position_request_pan -= self.__pan_step_value
                     
-                if self.__position_request.pan < self.__pan_min:
+                if self.__position_request_pan < self.__pan_min:
                     # Moved out of range, put back on min
-      	            self.__position_request.pan = self.__pan_min 
+      	            self.__position_request_pan = self.__pan_min 
       	            
       	if all_areas_scanned == True:
       	    # Reset camera/head position to default values                
-            self.__position_request.pan = self.__camera_default_pan_position
-            self.__position_request.tilt = self.__camera_default_tilt_position                    
-       
-        return all_areas_scanned, self.__position_request
+            self.__position_request_pan = self.__camera_default_pan_position
+            self.__position_request_tilt = self.__camera_default_tilt_position                      
+     
+        return all_areas_scanned, self.__position_request_pan, self.__position_request_tilt
         
     def CameraManualMove(self, direction):
-        relative_request = pan_tilt(0, 0)
+        relative_request_pan = 0.0
+        relative_request_tilt = 0.0
         # Check for up command
-        if 'u' in direction:
-            relative_request.tilt = self.__manual_tilt_step_value
-            if (self.__position_request.tilt + relative_request.tilt) > self.__tilt_max:
+        if 'd' in direction:
+            relative_request_tilt = self.__manual_tilt_step_value
+            if (self.__position_request_tilt + relative_request_tilt) > self.__tilt_max:
                 # Would move out of range so no change                   
-	            relative_request.tilt = 0
+	            relative_request_tilt = 0
             else:
                 # Keep track
-                self.__position_request.tilt += relative_request.tilt
+                self.__position_request_tilt += relative_request_tilt
                     
         # Check for down command
-        if 'd' in direction:
-            relative_request.tilt = -(self.__manual_tilt_step_value)
-            if (self.__position_request.tilt + relative_request.tilt) < self.__tilt_min:
+        if 'u' in direction:
+            relative_request_tilt = -(self.__manual_tilt_step_value)
+            if (self.__position_request_tilt + relative_request_tilt) < self.__tilt_min:
                 # Would move out of range so no change                 
-	            relative_request.tilt = 0
+	            relative_request_tilt = 0
             else:
                 # Keep track
-                self.__position_request.tilt += relative_request.tilt
+                self.__position_request_tilt += relative_request_tilt
             
         # Check for left commnand
         if 'l' in direction:
-            relative_request.pan = self.__manual_pan_step_value
-            if (self.__position_request.pan + relative_request.pan) > self.__pan_max:
+            relative_request_pan = self.__manual_pan_step_value
+            if (self.__position_request_pan + relative_request_pan) > self.__pan_max:
                 # Would move out of range so no change                   
-	            relative_request.pan = 0
+	            relative_request_pan = 0
             else:
                 # Keep track
-                self.__position_request.pan += relative_request.pan
+                self.__position_request_pan += relative_request_pan
             
         # Check for right command
         if 'r' in direction:
 
-            relative_request.pan = -(self.__manual_pan_step_value)
-            if (self.__position_request.pan + relative_request.pan) < self.__pan_min:
+            relative_request_pan = -(self.__manual_pan_step_value)
+            if (self.__position_request_pan + relative_request_pan) < self.__pan_min:
                 # Would move out of range so no change                   
-	            relative_request.pan = 0
+	            relative_request_pan = 0
             else:
                 # Keep track
-                self.__position_request.pan += relative_request.pan            
+                self.__position_request_pan += relative_request_pan            
             
-        return relative_request
+        return relative_request_pan, relative_request_tilt
     
                           
 # The PREPARE state
 class Prepare(State):
     def __init__(self, helper_obj):
-        State.__init__(self, outcomes=['mission2','mission3','done_task','head_default','move_head'], 
+        State.__init__(self, outcomes=['mission2','done_task','head_default','move_head'], 
                        input_keys=['mission'],
                        output_keys=['mission_data','start','user_data_absolute','user_data_pan','user_data_tilt'])
         self.__helper_obj = helper_obj
@@ -191,12 +188,6 @@ class Prepare(State):
             # other parameters with this mission request
             userdata.start = True
             retVal = 'mission2'
-        elif parameters[0] == 'M3':
-            # Mission 3 is to scan for the given object
-            # parameters[1] will contain the object to detect
-            userdata.mission_data = parameters[1]
-            userdata.start = True
-            retVal = 'mission3'
         elif parameters[0] == 'J1':
             # Simple Job 1 is play a supplied wav file and move the face lips           
             # Publish topic for speech wav and robot face animation
@@ -214,11 +205,11 @@ class Prepare(State):
             if 'c' in parameters[1]:
                 retVal = 'head_default'
             else:                
-                relative_request = self.__helper_obj.CameraManualMove(parameters[1]+parameters[2])                
+                relative_request_pan, relative_request_tilt = self.__helper_obj.CameraManualMove(parameters[1]+parameters[2])                
                 # Set up user data that will be used for goal in next state
                 userdata.user_data_absolute = False # This will be a relative move
-                userdata.user_data_pan = relative_request.pan
-                userdata.user_data_tilt = relative_request.tilt
+                userdata.user_data_pan = relative_request_pan
+                userdata.user_data_tilt = relative_request_tilt
                 retVal = 'move_head'
 
         return retVal
@@ -248,12 +239,6 @@ class RodneyMissionsNode:
         # Create an instance of the missions helper class
         self.__missions_helper = MissionsHelper()
         
-        # Create an instance of the helper class for greeting
-        self.__greet_helper = missions_lib.GreetingHelper()
-        
-        # Create an instance of the helper class for object search
-        self.__os_helper = missions_lib.ObjectSearchHelper()
-        
         # ------------------------- Top level state machine -------------------------
         # Create top level state machine
         self.__sm = StateMachine(outcomes=['preempted','aborted'],
@@ -271,7 +256,7 @@ class RodneyMissionsNode:
             # Add state to prepare the mission            
             StateMachine.add('PREPARE',
                              Prepare(self.__missions_helper),                             
-                             transitions={'mission2':'MISSION2','mission3':'MISSION3',
+                             transitions={'mission2':'MISSION2',
                                           'done_task':'WAITING','head_default':'DEFAULT_HEAD_POSITION',
                                           'move_head':'MOVE_HEAD'})
                              
@@ -280,13 +265,12 @@ class RodneyMissionsNode:
                              Report(),
                              transitions={'success':'DEFAULT_HEAD_POSITION'})
 
-            # Set up action goal for deafult head position
-            #default_position = pan_tilt()
-            default_position = self.__missions_helper.CameraDefaultPos()
+            # Set up action goal for deafult head position            
+            default_position_pan, default_position_tilt = self.__missions_helper.CameraDefaultPos()
             head_goal = point_headGoal()
             head_goal.absolute = True
-            head_goal.pan = default_position.pan
-            head_goal.tilt = default_position.tilt
+            head_goal.pan = default_position_pan
+            head_goal.tilt = default_position_tilt
 
             # Add the default camera position state. Which moves the head to the default position
             StateMachine.add('DEFAULT_HEAD_POSITION',
@@ -304,117 +288,15 @@ class RodneyMissionsNode:
                              transitions={'succeeded':'WAITING', 'preempted':'WAITING', 'aborted':'aborted'},
                              remapping={'absolute':'user_data_absolute', 'pan':'user_data_pan', 'tilt':'user_data_tilt'}) 
 
+            # ------------------------- Sub State machine for mission 2 ---------------------
+            # Create a sub state machine for mission 2 - face detection and greeting
+            self.__sm_mission2 = missions_lib.Mission2StateMachine(self.__missions_helper)
 
-            # ------------------------- State machine for mission 2 -------------------------
-            # Create a sub state machine for mission 2 - greeting
-            self.__sm_mission2 = StateMachine(outcomes=['complete','preempted','aborted'],
-                                              input_keys=['start'])
-
-            with self.__sm_mission2:
-                # This state will calculate the next head/camera position
-                StateMachine.add('PREPARE_FOR_MOVEMENT_GRT',
-                                 missions_lib.PrepareMovementGeeting(self.__missions_helper),
-                                 transitions={'complete':'GREETING',
-                                              'move':'MOVE_HEAD_GRT'},
-                                 remapping={'start_in':'start','start_out':'start'})
- 
-                # This state will call the action to move the head/camera and on completion request the face recognition
-                # operation on the next image
-                StateMachine.add('MOVE_HEAD_GRT',
-                                 SimpleActionState('head_control_node',
-                                                   point_headAction,
-                                                   result_cb = self.__greet_helper.MovementCompleteCB,
-                                                   goal_slots=['absolute','pan','tilt']),
-                                 transitions={'succeeded':'WAITING_FOR_ANALYSIS_GRT','preempted':'preempted','aborted':'aborted'},
-                                 remapping={'absolute':'user_data_absolute','pan':'user_data_pan','tilt':'user_data_tilt'})                                          
-
-                # This state will wait until face recognition is run on the
-                # image and the results returned
-                StateMachine.add('WAITING_FOR_ANALYSIS_GRT',
-                                 MonitorState('/face_recognition_node/result',
-                                              face_recognition,
-                                              self.__greet_helper.AnalysisCompleteGrt,
-                                              input_keys = ['seen_dict_in'],
-                                              output_keys = ['seen_dict_out']),
-                                 remapping={'seen_dict_in':'seen_dict',
-                                            'seen_dict_out':'seen_dict'},
-                                 transitions={'valid':'WAITING_FOR_ANALYSIS_GRT',
-                                              'invalid':'PREPARE_FOR_MOVEMENT_GRT',
-                                              'preempted':'preempted'}) 
-                                                                      
-                StateMachine.add('GREETING',
-                                 missions_lib.Greeting(self.__missions_helper),
-                                 transitions={'complete':'complete'})
-                                             
-            # Now add the sub state machine for mission 3 to the top level one
+            # Now add the sub state machine for mission 2 to the top level one
             StateMachine.add('MISSION2', 
                              self.__sm_mission2, 
-                             transitions={'complete':'REPORT','preempted':'REPORT','aborted':'aborted'})
-                                         
-            # ------------------------- State machine for mission 3 -------------------------        
-            # Create a sub state machine for mission 3 - object search
-            self.__sm_mission3 = StateMachine(outcomes=['complete', 'preempted','aborted'],
-                                              input_keys=['mission_data','start'])
-            
-            with self.__sm_mission3:
-                # This state will calculate the next head/camera position
-                StateMachine.add('PREPARE_FOR_MOVEMENT_OS',
-                                 missions_lib.PrepareMovementObjectSearch(self.__missions_helper),
-                                 transitions={'complete':'complete',
-                                              'move':'MOVE_HEAD_OS'},
-                                 remapping={'start_in':'start','start_out':'start'})
-                                 
-                # This state will call the action to move the head/camera and on completion request the object detection
-                # operation on the next image image
-                StateMachine.add('MOVE_HEAD_OS',
-                                 SimpleActionState('head_control_node',
-                                                   point_headAction,         
-                                                   result_cb = self.__os_helper.MovementCompleteCB,
-                                                   goal_slots=['absolute','pan','tilt']),
-                                 transitions={'succeeded':'WAITING_FOR_ANALYSIS_OS','preempted':'preempted','aborted':'aborted'},
-                                 remapping={'absolute':'user_data_absolute','pan':'user_data_pan','tilt':'user_data_tilt'}) 
-                
-                # This state will wait until object detection is run on an
-                # image and the results returned
-                StateMachine.add('WAITING_FOR_ANALYSIS_OS',
-                                 MonitorState('/tf_object_detection_node/result',
-                                 detection_results,
-                                 self.__os_helper.AnalysisCompleteOS,
-                                 input_keys = ['mission_data'],
-                                 output_keys = ['object_detected']),
-                                 transitions={'valid':'WAITING_FOR_ANALYSIS_OS', 
-                                              'invalid':'CHOICE_OS',
-                                              'preempted':'preempted'}) 
-                
-                # This state will will make the decision which state to move to
-                # next. If the object in question was detected then it will move
-                # to a state that waits for user interaction. If the object was
-                # not detected then it will move to the state to move the camera.
-                # This seems a bit of an overkill but the MonitorState before it
-                # can only provide the three possible outcomes 'valid', 'invalid'
-                # or 'preempted'. This state will request speech to state the
-                # object was found.
-                StateMachine.add('CHOICE_OS',
-                                 missions_lib.ChoiceObjectSearch(self.__missions_helper),           
-                                 transitions= {'move_camera':'PREPARE_FOR_MOVEMENT_OS',
-                                               'await_user':'WAIT_FOR_USER_OS'})
-                
-                # This state will wait for a message indicating the user wishes
-                # to move on. It will allow the user to examine the camera feed
-                # once an object was detected
-                StateMachine.add('WAIT_FOR_USER_OS',
-                                 MonitorState('/missions/acknowledge',
-                                 Empty,
-                                 self.__os_helper.UserAckOS),
-                                 transitions={'valid':'WAIT_FOR_USER_OS',
-                                              'invalid':'PREPARE_FOR_MOVEMENT_OS',
-                                              'preempted':'preempted'})
-                                                            
-            # Now add the sub state machine for mission 3 to the top level one
-            StateMachine.add('MISSION3', 
-                             self.__sm_mission3, 
-                             transitions={'complete':'REPORT', 'preempted':'REPORT','aborted':'aborted'}) 
-
+                             transitions={'complete':'REPORT','preempted':'REPORT','aborted':'aborted'})                                        
+            # -------------------------------------------------------------------------------
             
         # Create and start the introspective server so that we can use smach_viewer
         sis = IntrospectionServer('server_name', self.__sm, '/SM_ROOT')
@@ -440,10 +322,7 @@ class RodneyMissionsNode:
     def CancelCallback(self, data):
         # If a sub statemachine for a mission is running then request it be preempted
         if self.__sm_mission2.is_running():
-            self.__sm_mission2.request_preempt()
-        elif self.__sm_mission3.is_running():
-           
-            self.__sm_mission3.request_preempt()        
+            self.__sm_mission2.request_preempt()        
         
     def ShutdownCallback(self):        
         self.__sm.request_preempt()